btree: move elements up from a leaf

+++ src/btree.cpp
@@ -1,4 +1,4 @@
 #include "btree.h"
 
-int Traversable::deepest = 0;
-int Traversable::shallowest = MAX_INT;
+size_t Traversable::deepest = 0;
+size_t Traversable::shallowest = MAX_INT;
